REFACTOR/ProductService - 패키지, 클래스 이름 변경
